Update Plugin for 2023.3-EAP1 & Improve JDK Compatibility
Updated plugin version to 2023.3-EAP1 in gradle.properties and Plugin.props for compatibility with the newly released 2023.3 EAP. Readme.md now contains instructions on how to run the plugin on Windows and macOS. Gradle version updated to 8.2.1 and Kotlin JVM version updated to 1.9.0 in build.gradle to stay up-to-date. JDK compatibility target was set to 17 in build.gradle and JavaToolInstaller was added in build.yaml to handle Java 17. Changelog.md was updated to track these updates.